requirements.txt
# AWS
boto3==1.28.44
botocore==1.31.44

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0

# Data Processing
pandas==2.0.3
numpy==1.25.2
pyarrow==12.0.1

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7

# Utils
python-dotenv==1.0.0
pytest==7.4.0
logging==0.4.9.6